refactor: move generateGradient from PokemonCard component's methods to project utils
