+++ src/components/PokemonCard.tsx
@@ -1,5 +1,6 @@
 import { Component } from 'react';
 import { Pokemon } from '../types/type';
+import gradientGenerator from '../utils/gradientGenerator';
 
 interface PokemonCardProps extends Pokemon {}
 
@@ -8,22 +9,9 @@ export default class PokemonCard extends Component<PokemonCardProps> {
     super(props);
   }
 
-  generateGradient(id: number) {
-    const gradients = [
-      'bg-gradient-to-r from-pink-500 to-yellow-500',
-      'bg-gradient-to-r from-green-400 to-blue-500',
-      'bg-gradient-to-r from-purple-400 to-pink-500',
-      'bg-gradient-to-r from-indigo-500 to-purple-500',
-      'bg-gradient-to-r from-teal-400 to-blue-400',
-    ];
-
-    const index = id % gradients.length;
-    return gradients[index];
-  }
-
   render() {
     const { id, name, image, type } = this.props;
-    const gradientClass = this.generateGradient(id);
+    const gradientClass = gradientGenerator.generateGradient(id);
 
     return (
       <div
